% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_qpcr.R
\name{plot_qpcr}
\alias{plot_qpcr}
\title{Plot normalized qPCR data in a nice format}
\usage{
plot_qpcr(the_data, norm)
}
\arguments{
\item{the_data}{A data frame containing normalized qPCR data. If read_qpcr is used this would be the second
element of the returned list. Make sure the data frame contains at least the columns "Sample" and "Genotype".}

\item{norm}{Reference gene to use}
}
\value{
A ggplot2 object
}
\description{
Takes normalized qPCR data and turns it into a ggplot2 object for visualization
}
\examples{
\dontrun{
plot_qpcr(my_pcrdata[[2]], norm = Gapdh)
}
}
