\name{qPCRSet-class}
\docType{class}
\alias{qPCRSet-class}
\alias{class:qPCRSet}
\alias{qPCRSet}
\alias{qSet}

\title{
  Class to Contain and Describe normalise qPCR Data
}

\description{
  Container for normalised qPCR Data
}
\section{Creating Objects}{

  \code{new("qPCRSet")}

}
%\section{Slots}{
%}
%\section{Methods}{
%}

\author{James Perkins}

%\examples{
%}

\keyword{classes}

