% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qpcr.R
\name{read_qpcr}
\alias{read_qpcr}
\title{Read and process qPCR data}
\usage{
read_qpcr(datafolder, metadatafile = "./metadata.csv", norm)
}
\arguments{
\item{datafolder}{The directory, which contains one or multiple Biorad
CSVs. Important: all CSV files in that location will be read and cause
problems if they are not in the right format!}

\item{metadatafile}{A CSV file that contains sample metadata. At least a
"Sample" and"Genotype" field have to be in that file}

\item{norm}{Reference gene to use}
}
\value{
A list of two data frames: The first one contains the data
  collapsed by technical replicates and contains quality control indicators.
  The second one contains the 2^-ddCt calculations.
}
\description{
Reads an exported Bio-Rad qPCR result file, calculates quality
  controls summaries and performs 2^-ddCt calculations.
}
\examples{
\dontrun{
read_qpcr("./raw_data/", metadatafile = "./metadata.csv", norm = Gapdh)
}
}
