\name{ReadTaq}
\alias{ReadTaq}
\title{ Read Taqman qPCR data }
\description{
  The function reads Taqman RT-qPCR data.
}
\usage{
ReadTaq(filename = character())
}
\arguments{
  \item{filename}{ character, name of Taqman data file.}
}
\details{
  Reads a single Taqman data file.
}
\value{Object of class \code{"ExpressionSet"}.}
%\references{ ~put references to the literature/web site here ~ }
\author{ James Perkins \email{jperkins@biochem.ucl.ac.uk} }
%\note{}
\seealso{ \code{\link[Biobase]{ExpressionSet-class}} }
%\examples{
%}
\keyword{classes}
