\name{qPCRBatch-class}
\docType{class}
\alias{qPCRBatch-class}
\alias{class:qPCRBatch}
\alias{qPCRBatch}
\alias{qSet}
\alias{exprs.well.order}
\alias{exprs.well.order<-}
\alias{exprs.well.order,qPCRBatch-method}
\alias{exprs.well.order<-,qPCRBatch-method}
\title{
  Class to Contain and Describe normalise qPCR Data
}

\description{
  Container for normalised qPCR Data
}
\section{Creating Objects}{
  \code{new("qPCRBatch")}
}
%\section{Slots}{
%}
\section{Methods}{
  \describe{
    \item{exprs.well.order}{(qPCRBatch): ... }
    \item{exprs.well.order<-}{(qPCRBatch): ... }
  }
}

\author{James Perkins}

%\examples{
%}

\keyword{classes}

