\name{qPCRSet-class}
\docType{class}
\alias{qPCRSet-class}
\alias{class:qPCRSet}
\alias{qPCRSet}
\alias{qSet}
\alias{exprs.well.order}
\alias{exprs.well.order<-}
\alias{exprs.well.order,qPCRSet-method}
\alias{exprs.well.order<-,qPCRSet-method}
\title{
  Class to Contain and Describe normalise qPCR Data
}

\description{
  Container for normalised qPCR Data
}
\section{Creating Objects}{
  \code{new("qPCRSet")}
}
%\section{Slots}{
%}
\section{Methods}{
  \describe{
    \item{exprs.well.order}{(qPCRSet): ... }
    \item{exprs.well.order<-}{(qPCRSet): ... }
  }
}

\author{James Perkins}

%\examples{
%}

\keyword{classes}

