\name{ReadqPCR-package}
\alias{ReadqPCR-package}
\alias{ReadqPCR}
\docType{package}
\title{Read qPCR data}
\description{
The package provides functions to read raw RT-qPCR data of different platforms.
}
\details{
\tabular{ll}{
Package: \tab ReadqPCR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-01-22\cr
Depends: \tab R(>= 2.10.0), Biobase, ReadqPCR, methods, RColorBrewer\cr
License: \tab LGPL-3\cr
LazyLoad: \tab yes\cr
}
library(ReadqPCR)
}
\author{
James Perkins <jperkins@biochem.ucl.ac.uk>\cr
Matthias Kohl <Matthias.Kohl@stamats.de>

Maintainer: James Perkins <jperkins@biochem.ucl.ac.uk>
}
%\references{
%~~ Literature or other references for background information ~~
%}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
%~~ the R documentation directory ~~
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
\examples{
## some examples are given in the vignette
\dontrun{
library(ReadqPCR)
vignette("ReadqPCR")
}
}
\keyword{ package }
